Rename "nbgrader assign" to "nbgrader generate_assignment"
