+++ nbgrader/converters/__init__.py
@@ -1,5 +1,6 @@
 from .base import BaseConverter, NbGraderException
 from .assign import Assign
+from .generate_assignment import GenerateAssignment
 from .autograde import Autograde
 from .feedback import Feedback
 
@@ -7,6 +8,7 @@ __all__ = [
     "BaseConverter",
     "NbGraderException",
     "Assign",
+    "GenerateAssignment",
     "Autograde",
     "Feedback"
 ]
