+++ nbgrader/converters/base.py
@@ -37,7 +37,7 @@ class BaseConverter(LoggingConfigurable):
         help=dedent(
             """
             Permissions to set on files output by nbgrader. The default is generally
-            read-only (444), with the exception of nbgrader assign and nbgrader feedback,
+            read-only (444), with the exception of nbgrader generate_assignment and nbgrader feedback,
             in which case the user also has write permission.
             """
         )
