+++ nbgrader/docs/source/user_guide/faq.rst
@@ -146,7 +146,7 @@ Windows
 ~~~~~~~
 
 The core pieces of nbgrader will also work on Windows: the "Create Assignment"
-extension, ``nbgrader assign``, ``nbgrader autograde``, ``nbgrader formgrade``,
+extension, ``nbgrader generate_assignment``, ``nbgrader autograde``, ``nbgrader formgrade``,
 ``nbgrader feedback``, ``nbgrader validate``, and ``nbgrader export``.
 
 However, the parts of nbgrader corresponding to file management (the
