+++ nbgrader/tests/nbextensions/test_assignment_list.py
@@ -195,7 +195,7 @@ def test_show_assignments_list(browser, port, class_files, tempdir):
     _wait(browser).until(EC.visibility_of_element_located((By.CSS_SELECTOR, "#submitted_assignments_list_placeholder")))
 
     # release an assignment
-    run_nbgrader(["assign", "Problem Set 1", "--force"])
+    run_nbgrader(["generate_assignment", "Problem Set 1", "--force"])
     run_nbgrader(["release", "Problem Set 1", "--course", "abc101", "--force"])
 
     # click the refresh button
@@ -215,7 +215,7 @@ def test_multiple_released_assignments(browser, port, class_files, tempdir):
     _wait_until_loaded(browser)
 
     # release another assignment
-    run_nbgrader(["assign", "ps.01", "--force"])
+    run_nbgrader(["generate_assignment", "ps.01", "--force"])
     run_nbgrader(["release", "ps.01", "--course", "xyz 200", "--force"])
 
     # click the refresh button
@@ -476,7 +476,7 @@ def test_missing_exchange(exchange, browser, port, class_files, tempdir):
     os.makedirs(exchange)
 
     # release an assignment
-    run_nbgrader(["assign", "Problem Set 1", "--force"])
+    run_nbgrader(["generate_assignment", "Problem Set 1", "--force"])
     run_nbgrader(["release", "Problem Set 1", "--course", "abc101", "--force"])
 
     # click the refresh button
